+++ pkg/apis/lbaas/v1/bind_types.go
@@ -1,10 +1,12 @@
 package v1
 
+import "go.anx.io/go-anxcloud/pkg/apis/internal/gs"
+
 // anxcloud:object:hooks=RequestBodyHook
 
 // Bind represents an LBaaS FrontendBind
 type Bind struct {
-	commonMethods
+	gs.GenericService
 	HasState
 
 	CustomerIdentifier string     `json:"customer_identifier,omitempty"`